rename from AddRequirementType to CreateRequirementType
