+++ src/Equinor.Procosys.Preservation.Command/RequirementTypeCommands/CreateRequirementType/CreateRequirementTypeCommandHandler.cs
@@ -5,22 +5,22 @@ using Equinor.Procosys.Preservation.Domain.AggregateModels.RequirementTypeAggreg
 using MediatR;
 using ServiceResult;
 
-namespace Equinor.Procosys.Preservation.Command.RequirementTypeCommands.AddRequirementType
+namespace Equinor.Procosys.Preservation.Command.RequirementTypeCommands.CreateRequirementType
 {
-    public class AddRequirementTypeCommandHandler : IRequestHandler<AddRequirementTypeCommand, Result<int>>
+    public class CreateRequirementTypeCommandHandler : IRequestHandler<CreateRequirementTypeCommand, Result<int>>
     {
         private readonly IRequirementTypeRepository _requirementTypeRepository;
         private readonly IUnitOfWork _unitOfWork;
         private readonly IPlantProvider _plantProvider;
 
-        public AddRequirementTypeCommandHandler(IRequirementTypeRepository modeRepository, IUnitOfWork unitOfWork, IPlantProvider plantProvider)
+        public CreateRequirementTypeCommandHandler(IRequirementTypeRepository modeRepository, IUnitOfWork unitOfWork, IPlantProvider plantProvider)
         {
             _requirementTypeRepository = modeRepository;
             _unitOfWork = unitOfWork;
             _plantProvider = plantProvider;
         }
 
-        public async Task<Result<int>> Handle(AddRequirementTypeCommand request, CancellationToken cancellationToken)
+        public async Task<Result<int>> Handle(CreateRequirementTypeCommand request, CancellationToken cancellationToken)
         {
             var newRequirementType = new RequirementType(_plantProvider.Plant, request.Code, request.Title, request.SortKey);
             _requirementTypeRepository.Add(newRequirementType);
